backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
pydantic==2.7.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
alembic==1.13.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
